[DROOLS-6641] Fix as per PR request
