+++ kie-pmml-trusty/kie-pmml-commons/src/main/java/org/kie/pmml/commons/Constants.java
@@ -52,7 +52,6 @@ public class Constants {
     public static final String REASON_CODE_ALGORITHM = "reasonCodeAlgorithm";
     public static final String EMPTY_LIST = "emptyList";
     public static final String AS_LIST = "asList";
-    public static final String DUMP_KIE_PMML_SOURCES = "dumpKiePmmlSources";
 
     private Constants() {
         // Avoid instantiation
